break up methods into modules
