+++ dependz.gemspec
@@ -6,7 +6,7 @@ Gem::Specification.new do |s|
   s.version     = '0.0.1'
   s.date        = '2020-09-17'
   s.summary     = 'ruby dependency management gem.'
-  s.description = "Store dependencies in a graph struct, and sort them in topological order. No dpends_by will come be before it's depends_on."
+  s.description = 'Topoligical sorting for dependent relations'
   s.authors     = ['Rundong Gao']
   s.email       = 'asphinx423@gmail.com'
   s.files       = ['lib/dependz.rb']
