+++ lib/dependz.rb
@@ -4,19 +4,13 @@ require 'rgl/adjacency'
 require 'rgl/topsort'
 
 require 'dependz/add'
+require 'dependz/base'
+require 'dependz/sort_item'
 
 module Dependz
-  class CircularDependenceError < StandardError; end
-
   class Client
+    include Base
     include Add
-
-    def initialize
-      @dag = RGL::DirectedAdjacencyGraph.new
-    end
-
-    def sort
-      @dag.topsort_iterator.to_a.reverse
-    end
+    include SortItem
   end
 end
